requirements.txt
# Django
Django==2.0.4
djangorestframework==3.8.2

# misc
requests==2.18.4

# API documentation
coreapi==2.3.3
Markdown==2.6.11
Pygments==2.2.0
